[Updaet] UI
==버그수정==
ㆍ2인 이상 참가했을 때, 아이템이 2배로 추가되던 현상 제거
ㆍ인벤토리에 이미 아이템이 있을 때, 아이템이 배수로 추가되던 현상제거

==수정==
ㆍGounded, Jump, AirJumpUp, AirJumping 등기존의 Mouse0을 입력던 코드에서 아이템을 사용할 경우를 예외처리

ㆍUseItem RPC처리

+++ Assets/Scripts/Item/ItemScriptable.cs
@@ -6,7 +6,7 @@ public class ItemScriptable : ScriptableObject
     public string itemName;
     public Sprite icon;
     public string description;
-    public int effectValue;
+    public float effectValue;
 
     // 나중에 분류 용도 (예: 회복, 공격 등)
     public ItemType type;
